backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0 
pydantic==2.11.7 
python-dotenv==1.1.1 
motor==3.7.1 
httpx==0.27.0
passlib[bcrypt]==1.7.4 
psutil==5.9.5
python-jose[cryptography]==3.3.0